0206/back/pyproject.toml
[project]
name = "back"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
]
